% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastMCD.R
\name{cstep}
\alias{cstep}
\title{Perform single iteration of C-step}
\usage{
cstep(X, T, S, h)
}
\arguments{
\item{X}{A 2D matrix}

\item{T}{A vector of the initial location estimate}

\item{S}{A vector of the initial scatter estimate}

\item{h}{An integer specifying the number of observations to use}
}
\value{
A list of set (H), scatter (S) and location (T)
}
\description{
Perform single iteration of C-step
}
\keyword{internal}
