% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastMCD.R
\name{draw_h}
\alias{draw_h}
\title{Randomly draw a subset of observations}
\usage{
draw_h(X, h)
}
\arguments{
\item{X}{A 2D matrix}

\item{h}{An integer specifying the number of observations to use}
}
\value{
A vector representing an h-length subset of X
}
\description{
Randomly draw a subset of observations
}
