% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastMCD.R
\name{fastMCD}
\alias{fastMCD}
\title{Estimate location and scatter using FAST-MCD algorithm}
\usage{
fastMCD(X, h = 0)
}
\arguments{
\item{X}{A 2D matrix to estimate location and scatter from}

\item{h}{An integer optionally specifying number of observations to use}
}
\value{
A list of estimated location (center) and scatter (cov)
}
\description{
Estimate location and scatter using FAST-MCD algorithm
}
\examples{
set.seed(51234)
S <- matrix(runif(5^2), 5)
S <- t(S) \%*\% S
X <- MASS::mvrnorm(700, mu = rep(0, 5), Sigma = S) # generate random matrix
outliers <- MASS::mvrnorm(140, mu = rep(5, 5), Sigma = S)
X[seq(1, 700, 5), ] <- outliers # set 20\% of observations to be outliers 
res <- fastMCD(X) # estimate location and scatter
}
