% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastMCD.R
\name{pick10}
\alias{pick10}
\title{Choose the 10 best estimates after iterating twice through initial sets}
\usage{
pick10(X, H_all, h)
}
\arguments{
\item{X}{A 2D matrix}

\item{H_all}{A 2D matrix where each row specifies a subset of observations}

\item{h}{An integer specifying number of observations to use}
}
\value{
A list of best sets (H), scatter (S) and location (T)
}
\description{
Choose the 10 best estimates after iterating twice through initial sets
}
