% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastMCD.R
\name{step_it}
\alias{step_it}
\title{Iterate through C-step}
\usage{
step_it(X, T, S, h, it = 0)
}
\arguments{
\item{X}{A 2D matrix}

\item{T}{A vector of the initial location estimate}

\item{S}{A vector of the initial scatter estimate}

\item{h}{An integer specifying the number of observations to use}

\item{it}{An optional integer specifying the number of C-steps to perform.
With it = 0, C-step will be performed until convergence}
}
\value{
A list of set (H), scatter (S) and location (T)
}
\description{
Iterate through C-step
}
