% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastMCD.R
\name{bigMCD}
\alias{bigMCD}
\title{Obtain unweighted estimates for data with > 600 observations}
\usage{
bigMCD(X, h, p, n)
}
\arguments{
\item{X}{A 2D matrix to estimate location and scatter from}

\item{h}{An integer specifying number of observations to use}

\item{p}{An integer specifying the number of columns in X}

\item{n}{An integer specifying the number of total observations}
}
\value{
A list of estimated location (center) and scatter (cov)
}
\description{
Obtain unweighted estimates for data with > 600 observations
}
